+++ libs/panic-abort/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "panic-abort"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+
+[lib]
+harness = false
+
+[lints]
+workspace = true
+
+[dependencies]
+panic-common.workspace = true
+log.workspace = true
