+++ libs/panic-common/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "panic"
+name = "panic-common"
 version.workspace = true
 edition.workspace = true
 authors.workspace = true
@@ -8,19 +8,11 @@ license.workspace = true
 [lints]
 workspace = true
 
-[lib]
-harness = false
-
 [dependencies]
-cfg-if.workspace = true
 sync.workspace = true
 log.workspace = true
-tls = { workspace = true, optional = true }
-unwind = { workspace = true, optional = true }
+cfg-if.workspace = true
 
 [target.'cfg(any(target_arch = "riscv64", target_arch = "riscv32"))'.dependencies]
 riscv.workspace = true
 
-[features]
-abort = []
-unwind = ["dep:tls", "dep:unwind"]
