--- libs/panic/src/unwind/mod.rs
@@ -1,81 +1,0 @@
-use crate::arch;
-use alloc::boxed::Box;
-use core::any::Any;
-use core::panic::PanicPayload;
-
-pub mod panic_count;
-
-/// Mirroring std, this is an unmangled function on which to slap
-/// yer breakpoints for backtracing panics.
-#[inline(never)]
-#[no_mangle]
-pub fn rust_panic(payload: &mut dyn PanicPayload) -> ! {
-    let code = unwind::panic_begin(unsafe { Box::from_raw(payload.take_box()) });
-
-    arch::exit(code);
-}
-
-#[allow(clippy::items_after_statements)]
-pub unsafe fn r#try<R, F: FnOnce() -> R>(f: F) -> Result<R, Box<dyn Any + Send>> {
-    use core::{intrinsics, mem::ManuallyDrop, ptr::addr_of_mut};
-
-    union Data<F, R> {
-        // when we start, this field holds the closure
-        f: ManuallyDrop<F>,
-        // when the closure completed successfully, this will hold the return
-        r: ManuallyDrop<R>,
-        // when the closure panicked this will hold the panic payload
-        p: ManuallyDrop<Box<dyn Any + Send>>,
-    }
-
-    #[inline]
-    fn do_call<F: FnOnce() -> R, R>(data: *mut u8) {
-        // SAFETY: this is the responsibility of the caller, see above.
-        unsafe {
-            let data = data.cast::<Data<F, R>>();
-            let data = &mut (*data);
-            let f = ManuallyDrop::take(&mut data.f);
-            data.r = ManuallyDrop::new(f());
-        }
-    }
-
-    #[inline]
-    #[rustc_nounwind] // `intrinsic::r#try` requires catch fn to be nounwind
-    fn do_catch<F: FnOnce() -> R, R>(data: *mut u8, payload: *mut u8) {
-        // SAFETY: this is the responsibility of the caller, see above.
-        //
-        // When `__rustc_panic_cleaner` is correctly implemented we can rely
-        // on `obj` being the correct thing to pass to `data.p` (after wrapping
-        // in `ManuallyDrop`).
-        unsafe {
-            let data = data.cast::<Data<F, R>>();
-            let data = &mut (*data);
-            let obj = cleanup(payload);
-            data.p = ManuallyDrop::new(obj);
-        }
-    }
-
-    #[cold]
-    unsafe fn cleanup(payload: *mut u8) -> Box<dyn Any + Send + 'static> {
-        // SAFETY: The whole unsafe block hinges on a correct implementation of
-        // the panic handler `__rust_panic_cleanup`. As such we can only
-        // assume it returns the correct thing for `Box::from_raw` to work
-        // without undefined behavior.
-        let obj = unsafe { unwind::panic_cleanup(payload) };
-        panic_count::decrease();
-        obj
-    }
-
-    let mut data = Data {
-        f: ManuallyDrop::new(f),
-    };
-    let data_ptr = addr_of_mut!(data).cast::<u8>();
-
-    unsafe {
-        if intrinsics::catch_unwind(do_call::<F, R>, data_ptr, do_catch::<F, R>) == 0 {
-            Ok(ManuallyDrop::into_inner(data.r))
-        } else {
-            Err(ManuallyDrop::into_inner(data.p))
-        }
-    }
-}
